requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
simplejson==3.19.2
urllib3==2.1.0
v4l2py==2.3.0
websocket-client==1.6.4
wsproto==1.2.0
yarl==1.9.3
